requirements.txt
# requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-pymongo==2.3.0
PyJWT==2.6.0